Firebase authentication.

+++ src/main/java/com/kamal/getmesocial/resources/MongoAlbumResource.java
@@ -1,9 +1,9 @@
 package com.kamal.getmesocial.resources;
 
-import com.kamal.getmesocial.model.MongoAlbum;
-import com.kamal.getmesocial.model.MongoComment;
-import com.kamal.getmesocial.model.MongoPhoto;
-import com.kamal.getmesocial.model.User;
+import com.google.firebase.auth.FirebaseAuthException;
+import com.kamal.getmesocial.exception.IdTokenException;
+import com.kamal.getmesocial.model.*;
+import com.kamal.getmesocial.service.FirebaseAdminService;
 import com.kamal.getmesocial.service.MongoAlbumService;
 import com.kamal.getmesocial.service.MongoCommentService;
 import com.kamal.getmesocial.service.UserService;
@@ -11,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,10 +21,19 @@ public class MongoAlbumResource {
     @Autowired
     private MongoAlbumService mongoAlbumService;
 
+    @Autowired
+    private FirebaseAdminService firebaseAdminService;
+
     @PostMapping("/album")
-    public MongoAlbum createAlbum(@RequestBody @Valid MongoAlbum mongoAlbum){
-        mongoAlbumService.createAlbum(mongoAlbum);
-        return mongoAlbum;
+    public MongoAlbum createAlbum(@RequestBody @Valid MongoAlbum mongoAlbum, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException, IdTokenException {
+        FirebaseUser falbum = firebaseAdminService.authentication(idToken);
+        if(falbum != null){
+            mongoAlbumService.createAlbum(mongoAlbum);
+            return mongoAlbum;
+        }else {
+            return null;
+        }
+
     }
     @GetMapping("/albums")
     public List<MongoAlbum> getAllAlbums() {
@@ -36,13 +46,23 @@ public class MongoAlbumResource {
 
     }
     @PutMapping("/album")
-    public MongoAlbum updateAlbum(@RequestBody MongoAlbum mongoAlbum){
-        return mongoAlbumService.updateAlbum(mongoAlbum);
+    public MongoAlbum updateAlbum(@RequestBody MongoAlbum mongoAlbum, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException, IdTokenException {
+        FirebaseUser falbum = firebaseAdminService.authentication(idToken);
+        if(falbum != null){
+            return mongoAlbumService.updateAlbum(mongoAlbum);
+        }else {
+            return null;
+        }
+
 
     }
     @DeleteMapping("/album")
-    public void deleteMongoAlbum(@RequestParam(name = "id") String albumId) {
-        mongoAlbumService.deleteAlbum(albumId);
+    public void deleteMongoAlbum(@RequestParam(name = "id") String albumId, @RequestHeader(name = "idToken") String idToken) throws IOException, FirebaseAuthException, IdTokenException {
+        FirebaseUser falbum = firebaseAdminService.authentication(idToken);
+        if(falbum != null){
+            mongoAlbumService.deleteAlbum(albumId);
+        }
+
 
     }
